works with a few warnings
