+++ packages/core/package.json
@@ -9,8 +9,8 @@
   "scripts": {
     "clean": "rimraf dist",
     "compile": "tsc",
-    "build": "rollup --config rollup.config.js",
-    "build:clean": "npm run clean && npm run compile && rollup --config rollup.config.js",
+    "build": "rollup --config rollup.config.mjs",
+    "build:clean": "npm run clean && npm run compile && rollup --config rollup.config.mjs",
     "build:watch": "npm run compile && npm run build -- --watch",
     "test": "jest unit",
     "test:watch": "jest unit --watch",
@@ -29,8 +29,8 @@
   },
   "devDependencies": {
     "@babel/preset-typescript": "^7.14.5",
-    "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-babel": "^6.0.4",
+    "@rollup/plugin-json": "^6.1.0",
     "@rollup/plugin-typescript": "^8.2.1",
     "@types/blueimp-md5": "^2.18.0",
     "@types/jest": "^27.0.6",
@@ -40,10 +40,12 @@
     "@types/uuid": "^8.3.1",
     "jest": "^27.0.6",
     "rimraf": "^3.0.2",
-    "rollup": "^2.52.7",
-    "rollup-plugin-dts": "^3.0.2",
+    "rollup": "^4.27.3",
+    "rollup-plugin-dts": "^6.1.1",
     "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-swc3": "^0.12.1",
     "rollup-plugin-terser": "^7.0.2",
+    "rollup-swc-preserve-directives": "^0.7.0",
     "ts-jest": "^27.0.3",
     "typescript": "^5.6.3"
   },
