+++ packages/core/rollup.config.mjs
@@ -1,9 +1,9 @@
 import babel from "@rollup/plugin-babel";
-import typescript from "@rollup/plugin-typescript";
 import json from "@rollup/plugin-json";
-import { terser } from "rollup-plugin-terser";
 import dts from "rollup-plugin-dts";
 import peerDepsExternal from "rollup-plugin-peer-deps-external";
+import { swc } from "rollup-plugin-swc3";
+import swcPreserveDirectives from "rollup-swc-preserve-directives";
 
 const globals = {
   react: "React",
@@ -33,24 +33,15 @@ const config = [
       {
         file: "dist/index.js",
         format: "cjs",
-      },
-      {
-        file: "dist/index.min.js",
-        format: "iife",
-        name: "gffCore",
-        plugins: [terser()],
         globals,
-      },
-      {
-        file: "dist/index.umd.js",
-        format: "umd",
-        name: "gffCore",
-        globals,
+        sourcemap: true,
       },
       {
         file: "dist/index.esm.js",
         format: "esm",
         name: "gffCore",
+        globals,
+        sourcemap: true,
       },
     ],
     external: [
@@ -68,8 +59,15 @@ const config = [
     ],
     plugins: [
       peerDepsExternal(),
-      typescript(),
       json(),
+      swc({
+        sourceMaps: true,
+        include: /\.[mc]?[jt]sx?$/,
+        exclude: /node_modules/,
+        tsconfig: "tsconfig.json",
+        jsc: {},
+      }),
+      swcPreserveDirectives(),
       babel({
         presets: ["@babel/preset-react"],
         plugins: ["@emotion"],
