+++ packages/core/src/features/cohort/tests/availableCohortSlice.unit.test.ts
@@ -15,13 +15,13 @@ import {
   selectCohortById,
   selectMultipleCohortsById,
 } from "../availableCohortsSlice";
-import { NullCountsData } from "../cohortCountsQuery";
 import * as cohortSlice from "../availableCohortsSlice";
 import { Dictionary, EntityState } from "@reduxjs/toolkit";
 import { MOCK_COHORTS } from "./mockData";
 import { FilterSet } from "../filters";
 import { getInitialCoreState } from "src/store.unit.test";
 import { DataStatus } from "src/dataAccess";
+import { NullCountsData } from "../type";
 
 const state = getInitialCoreState();
