+++ packages/core/src/features/tables/index.ts
@@ -1,3 +1,3 @@
-import { useGetGeneTableQuery, TableData } from "./table";
+import { useGetGeneTableQuery, type TableData } from "./table";
 
 export { useGetGeneTableQuery, TableData };
