+++ packages/core/src/index.ts
@@ -59,4 +59,4 @@ export * from "./features/sets";
 export * from "./features/bodyplot";
 export * from "./features/setOperations";
 export * from "./features/counts";
-export { CoreState } from "./reducers";
+export { type CoreState } from "./reducers";
