+++ packages/portal-components/package.json
@@ -9,8 +9,8 @@
   "scripts": {
     "clean": "rimraf dist",
     "compile": "tsc",
-    "build": "rollup --config rollup.config.js",
-    "build:clean": "npm run clean && npm run compile && rollup --config rollup.config.js",
+    "build": "rollup --config rollup.config.mjs",
+    "build:clean": "npm run clean && npm run compile && rollup --config rollup.config.mjs",
     "build:watch": "npm run compile && npm run build -- --watch",
     "test": "jest unit",
     "test:watch": "jest unit --watch",
@@ -26,19 +26,23 @@
     "tailwindcss": "^3.2.4"
   },
   "devDependencies": {
-    "@babel/preset-typescript": "^7.14.5",
-    "@rollup/plugin-babel": "^5.3.0",
+    "@babel/preset-typescript": "^7.26.0",
+    "@rollup/plugin-babel": "^6.0.4",
     "@rollup/plugin-node-resolve": "^15.3.0",
-    "@rollup/plugin-typescript": "^8.2.1",
+    "@rollup/plugin-typescript": "^12.1.1",
+    "@types/lodash": "4.17.13",
+    "@types/node": "22.9.1",
     "@types/react": "^18.3.3",
     "autoprefixer": "^10.4.20",
     "postcss": "^8.4.47",
-    "rollup": "^2.52.7",
+    "rollup": "^4.27.3",
     "rollup-plugin-dts": "^3.0.2",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-swc3": "^0.12.1",
+    "rollup-swc-preserve-directives": "^0.7.0",
     "tailwindcss": "^3.4.14",
-    "typescript": "~4.6.4"
+    "typescript": "^5.6.3"
   },
   "dependencies": {
     "rollup-plugin-postcss": "^4.0.2"
